58 factor out excel downloading logic (#82)
* Factor out excel downloading logic [ #58 ]

* Only export necessary columns in assignments

* use Pick instead of Omit
